refactored code for npm testing.
